Configured gradle build to run code, as well as general package housekeeping
