+++ src/main/java/seedu/financeit/utils/CommandPacket.java
@@ -1,4 +1,4 @@
-package utils; /**
+package seedu.financeit.utils; /**
  * Organises data from input string into distinct data types and structure.
  * For scalability purposes, a hash table is implemented to map param type to its corresponding params.
  */
