+++ src/main/java/seedu/financeit/utils/FiniteStateMachine.java
@@ -1,4 +1,4 @@
-package utils;
+package seedu.financeit.utils;
 
 public class FiniteStateMachine {
     public enum State {
