+++ src/main/java/seedu/financeit/utils/RegexMatcher.java
@@ -1,4 +1,4 @@
-package utils;
+package seedu.financeit.utils;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
